Got bones showing

+++ FAST_RTA_SOLUTION/FAST_RTA_SOLUTION/Animation.cpp
@@ -18,7 +18,7 @@ const KeyFrame * Animation::GetFrame(int _index) const
 
 unsigned int Animation::GetNumKeyFrames() const
 {
-	return m_keyFrames.size();
+	return (unsigned int)m_keyFrames.size();
 }
 
 unsigned int Animation::GetNumBones() const
@@ -30,6 +30,6 @@ void Animation::Init(std::vector<KeyFrame> &_keyFrames, ANIM_TYPE _animType)
 {
 	m_keyFrames = _keyFrames;
 	m_animType = _animType;
-	m_numBones = m_keyFrames[0].m_bones.size();
+	m_numBones = (unsigned int)m_keyFrames[0].m_bones.size();
 	m_totaltime = m_keyFrames[m_keyFrames.size() - 1].m_time - m_keyFrames[0].m_time;
 }
